Links to logs should point to unified logs if enabled
