+++ apps/studio/components/interfaces/Observability/ObservabilityOverview.tsx
@@ -6,6 +6,7 @@ import { useRouter } from 'next/router'
 import { useCallback, useMemo, useState } from 'react'
 import { Badge, Button, Tooltip, TooltipContent, TooltipTrigger } from 'ui'
 
+import { useUnifiedLogsPreview } from '../App/FeaturePreview/FeaturePreviewContext'
 import { DatabaseInfrastructureSection } from './DatabaseInfrastructureSection'
 import { useObservabilityOverviewData } from './ObservabilityOverview.utils'
 import { ObservabilityOverviewFooter } from './ObservabilityOverviewFooter'
@@ -30,6 +31,7 @@ export const ObservabilityOverview = () => {
   const { data: organization } = useSelectedOrganizationQuery()
   const queryClient = useQueryClient()
 
+  const { isEnabled: isUnifiedLogsEnabled } = useUnifiedLogsPreview()
   const { projectStorageAll: storageSupported } = useIsFeatureEnabled(['project_storage:all'])
   const { isEnabled: isDataApiEnabled } = useIsDataApiEnabled({ projectRef })
 
@@ -79,7 +81,9 @@ export const ObservabilityOverview = () => {
         key: 'db' as const,
         name: 'Database',
         reportUrl: `/project/${projectRef}/observability/database`,
-        logsUrl: `/project/${projectRef}/logs/postgres-logs`,
+        logsUrl: isUnifiedLogsEnabled
+          ? `/project/${projectRef}/logs?filter=log_type:eq:postgres`
+          : `/project/${projectRef}/logs/postgres-logs`,
         enabled: true,
         hasReport: true,
       },
@@ -87,7 +91,9 @@ export const ObservabilityOverview = () => {
         key: 'postgrest' as const,
         name: 'PostgREST',
         reportUrl: `/project/${projectRef}/observability/postgrest`,
-        logsUrl: `/project/${projectRef}/logs/postgrest-logs`,
+        logsUrl: isUnifiedLogsEnabled
+          ? `/project/${projectRef}/logs?filter=log_type:eq:postgrest`
+          : `/project/${projectRef}/logs/postgrest-logs`,
         enabled: true,
         hasReport: true,
       },
@@ -95,7 +101,9 @@ export const ObservabilityOverview = () => {
         key: 'auth' as const,
         name: 'Auth',
         reportUrl: `/project/${projectRef}/observability/auth`,
-        logsUrl: `/project/${projectRef}/logs/auth-logs`,
+        logsUrl: isUnifiedLogsEnabled
+          ? `/project/${projectRef}/logs?filter=log_type:eq:auth`
+          : `/project/${projectRef}/logs/auth-logs`,
         enabled: true,
         hasReport: true,
       },
@@ -103,7 +111,9 @@ export const ObservabilityOverview = () => {
         key: 'functions' as const,
         name: 'Edge Functions',
         reportUrl: `/project/${projectRef}/observability/edge-functions`,
-        logsUrl: `/project/${projectRef}/logs/edge-functions-logs`,
+        logsUrl: isUnifiedLogsEnabled
+          ? `/project/${projectRef}/logs?filter=log_type:eq:edge+function`
+          : `/project/${projectRef}/logs/edge-functions-logs`,
         enabled: true,
         hasReport: true,
       },
@@ -111,7 +121,9 @@ export const ObservabilityOverview = () => {
         key: 'storage' as const,
         name: 'Storage',
         reportUrl: `/project/${projectRef}/observability/storage`,
-        logsUrl: `/project/${projectRef}/logs/storage-logs`,
+        logsUrl: isUnifiedLogsEnabled
+          ? `/project/${projectRef}/logs?filter=log_type:eq:storage`
+          : `/project/${projectRef}/logs/storage-logs`,
         enabled: storageSupported,
         hasReport: true,
       },
@@ -119,12 +131,14 @@ export const ObservabilityOverview = () => {
         key: 'realtime' as const,
         name: 'Realtime',
         reportUrl: `/project/${projectRef}/observability/realtime`,
-        logsUrl: `/project/${projectRef}/logs/realtime-logs`,
+        logsUrl: isUnifiedLogsEnabled
+          ? `/project/${projectRef}/logs?filter=log_type:eq:realtime`
+          : `/project/${projectRef}/logs/realtime-logs`,
         enabled: true,
         hasReport: true,
       },
     ],
-    [projectRef, storageSupported, isDataApiEnabled]
+    [projectRef, storageSupported, isDataApiEnabled, isUnifiedLogsEnabled]
   )
 
   const enabledServices = serviceBase.filter((s) => s.enabled)
